be-feat: refactor
