+++ repository/booking/__init__.py
@@ -1,5 +1,4 @@
 from .Order import OrderRepository
-from .Dish import DishRepository
 from .OrderItem import OrderItemRepository
-from .Table import TableRepository
+
