+++ repository/resources/Table.py
@@ -2,7 +2,7 @@ from sqlalchemy.ext.asyncio import AsyncSession
 from sqlalchemy import and_, delete, select, update
 from models import Table
 
-from schemas.booking import (
+from schemas.resources import (
     TableCreate,
     TableUpdate,
     TableFilter,
