+++ repository/resources/__init__.py
@@ -1,2 +1,4 @@
 from .Equipment import EquipmentRepository, EquipmentTypeRepository, EquipmentStatusRepository
 from .Ingredient import IngredientRepository, IngredientUnitRepository, IngredientHistoryRepository
+from .Table import TableRepository
+from .Dish import DishRepository
