+++ schemas/booking/__init__.py
@@ -1,5 +1,4 @@
-from .Dish import DishCreate, DishRead, DishUpdate, DishFilter
 from .OrderItem import OrderItemCreate, OrderItemRead, OrderItemUpdate, OrderItemFilter, OrderItemBase
 from .Order import OrderCreate, OrderRead, OrderUpdate, OrderFilter
-from .Table import TableCreate, TableRead, TableUpdate, TableFilter
+
