+++ schemas/resources/__init__.py
@@ -5,3 +5,6 @@ from .Equipment import (EquipmentCreate, EquipmentFilter, EquipmentUpdate,Equipm
 from .Ingredient import (IngredientCreate, IngredientUpdate, IngredientFilter, IngredientRead,
     IngredientUnitCreate, IngredientUnitUpdate, IngredientUnitFilter,IngredientUnitRead,
     IngredientHistoryCreate, IngredientHistoryRead, IngredientHistoryUpdate, IngredientHistoryFilter)
+
+from .Dish import (DishCreate, DishRead, DishUpdate, DishFilter)
+from .Table import (TableCreate, TableRead, TableUpdate, TableFilter)
